Render the intent under the correct token column
So I'm going to split out the intent into an instruction meta thing of
its own, later we'll expand on the instruction set but for now, I want
to dettach the parsing from the "execution" so i can render the intended
execution behind an intent line, say this 10 times fast. Basically, If
i'm buying token1 I want to see buying under it, and receiving under
token0, and vice versa.