A very basic REPL to close the feedback loop with the user, along with a refactor
Not so big of a refactor but definitely a lot more structure than
before. Report generation now sits in report_tables, the loop is simple,
we take the intent and generate the first report, prompt the user for
feedback, proceed, cancel, or change something, change will take a new
intent and re-generate the report table.

Pulled a few things I have for termbox from an old project, progress
spinner, blinking cursor