Shuffling around all the prose
Old README goes into devlog because that's what it was
In with the new README, bare and uninviting as it should be.

Rename docs.md to contribute.md, because "usage" docs should go into the
README, when this is ready for usage.

I was kiding about the uninviting part, FYI.