Reimplemented the data flow to reflect what was actually happening and support max out trx
So this is a major overhaul. The essence of which is simple:

Pool x Intent -> CPIntent -> Transaction.

At the point in time where we have a Pool and an Intent we'd only be
missing a constant pool curve resolved to the correct variables. The
intent when analyzed will give us a direction, which is all encapsulated
in a CPIntent, from there we have everything we need to construct the
correct type of transaction and provide the correct input needed.