Restructure and clean up the mess around CPIntents
CPIntents no longer store adhoc fields with no clear direction behind
what goes in and where. The whole point behind this which I introduced
in commit f3a63efc8d2dc573d4d9932548436eea0011e1fb was to redirect the
flow of data into something that made sense. The idea was Pool x Intent
-> CPIntent, which would capture a ready to compute curve reflecting the
intent we were given.

From there, I can generate a transaction to reflect this on chain, I can
render a quote table to inform the user. But the code was an absolute
mess, so, this introduces the following

- SwapLeg now reflects one side of the coin, and there's always going to
be two clear sides, an In token and an Out token.
- SwapAmounts represent what we've come to know in terms of values
- PoolAccounts is everything we know about the pool we're interacting
with now.

NewCPIntent will fill this structure out, apply slippage on the constant
product values, and now the flow looks like this pool+balances+intent ->
resolved swap.

Everything else basically follows suit. Much cleaner. Still need to
adjust the TUI